+++ requrements.txt
@@ -1,0 +1,2 @@
+beautifulsoup4==4.10.0
+requests==2.27.1
